+++ trunk/GLAndroid/src/com/gooble/logic/app/entity/Hintrelation.java
@@ -2,18 +2,18 @@ package com.gooble.logic.app.entity;
 
 public class Hintrelation extends Entity{
 
-   private Integer hintid;
-   private Integer relationid;
-   public Integer getHintid() {
+   private Long hintid;
+   private Long relationid;
+   public Long getHintid() {
       return hintid;
    }
-   public void setHintid(Integer hintid) {
+   public void setHintid(Long hintid) {
       this.hintid = hintid;
    }
-   public int getRelationid() {
+   public Long getRelationid() {
       return relationid;
    }
-   public void setRelationid(Integer relationid) {
+   public void setRelationid(Long relationid) {
       this.relationid = relationid;
    }
