+++ trunk/GLAndroid/src/com/gooble/logic/app/entity/Variable.java
@@ -3,17 +3,18 @@ package com.gooble.logic.app.entity;
 public class Variable extends Entity{
 
    private String name;
-   private Integer puzzleid;
+   private Long puzzleid;
+   
    public String getName() {
       return name;
    }
    public void setName(String name) {
       this.name = name;
    }
-   public Integer getPuzzleid() {
+   public Long getPuzzleid() {
       return puzzleid;
    }
-   public void setPuzzleid(Integer puzzleid) {
+   public void setPuzzleid(Long puzzleid) {
       this.puzzleid = puzzleid;
    }
