+++ trunk/GLAndroid/src/com/gooble/logic/app/entity/Variablevalue.java
@@ -3,17 +3,17 @@ package com.gooble.logic.app.entity;
 public class Variablevalue extends Entity{
 
    private String value;
-   private Integer variableid;
+   private Long variableid;
    public String getValue() {
       return value;
    }
    public void setValue(String value) {
       this.value = value;
    }
-   public Integer getVariableid() {
+   public Long getVariableid() {
       return variableid;
    }
-   public void setVariableid(Integer variableid) {
+   public void setVariableid(Long variableid) {
       this.variableid = variableid;
    }
